chore: Update npm dependencies
This commit updates the npm dependencies in the package.json file. It adds "better-sqlite3" version "^10.0.0" and updates "next" to version "14.2.3". It also updates devDependencies, including "eslint" to version "^8" and "eslint-config-next" to version "14.2.3". Additionally, it updates "typescript" to version "^5". These updates ensure that the project is using the latest stable versions of the dependencies.